Cleaned up main.js and refined dropdown functionality
