+++ index.html
@@ -34,9 +34,12 @@
             <span>Plan Your Next Trip:</span>
 
             <div id="states-parks" class="dropdowns">
-                <select name="dropState" id="dropState">
+                <div class="stateDropDown">
+                  <select name="dropState" id="dropState">
                 
-                </select>
+                  </select>
+                </div>
+
                 <div id="parkDropDown">
 
                 </div>
